Pull out metric middleware logic
